+++ app/models/index.ts
@@ -12,7 +12,7 @@ const basename = path.basename(__filename);
 const sequelize = new Sequelize(dbConfig.database, dbConfig.username, dbConfig.password, dbConfig);
 
 fs.readdirSync(__dirname)
-  .filter(file => file.indexOf('.') !== 0 && file !== basename && file.slice(-3) === '.js')
+  .filter(file => file.indexOf('.') !== 0 && file !== basename && file.slice(-3) === '.ts')
   .forEach(file => {
     const model = sequelize.import(path.join(__dirname, file));
     models[model.name] = model;
