+++ package.json
@@ -9,9 +9,9 @@
   "scripts": {
     "build": "tsc",
     "console": "node console.js",
-    "cover": "NODE_ENV=testing jest --coverage --runInBand --forceExit test/app",
-    "test": "NODE_ENV=testing jest dist/test/app --runInBand --forceExit",
-    "test-inspect": "NODE_ENV=testing node --inspect --debug-brk jest test/app.spec.js",
+    "cover": "NODE_ENV=testing jest --coverage --runInBand --forceExit",
+    "test": "NODE_ENV=testing jest --runInBand --forceExit",
+    "test-inspect": "NODE_ENV=testing node --inspect --debug-brk jest",
     "coveralls": "npm run cover -- --report lcovonly && cat ./coverage/lcov.info | coveralls",
     "eslint-check": "eslint --print-config .eslintrc.js --ignore-pattern ./.eslintrc.js | eslint-config-prettier-check",
     "lint": "eslint \"**/*.ts\" --ignore-pattern ./.eslintrc.js",
@@ -43,7 +43,7 @@
   },
   "jest": {
     "collectCoverageFrom": [
-      "**/*.js",
+      "**/*.ts",
       "!**/console.js",
       "!**/node_modules/**",
       "!**/build/**",
@@ -54,8 +54,14 @@
     ],
     "testEnvironment": "node",
     "transform": {
-      "^.+\\.js$": "babel-jest"
-    }
+      "^.+\\.ts?$": "ts-jest"
+    },
+    "testMatch": [
+      "**/test/**/*.spec.ts?(x)"
+    ],
+    "setupFilesAfterEnv": [
+      "<rootDir>/test/app.ts"
+    ]
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
@@ -81,12 +87,14 @@
     "@types/body-parser": "^1.17.0",
     "@types/cors": "^2.8.5",
     "@types/express": "^4.17.0",
+    "@types/jest": "^24.0.18",
     "@types/jwt-simple": "^0.5.33",
     "@types/morgan": "^1.7.35",
     "@types/node": "^12.6.2",
     "@types/pg": "^7.4.14",
     "@types/request-promise": "^4.1.44",
     "@types/sequelize": "^4.28.3",
+    "@types/supertest": "^2.0.8",
     "@types/umzug": "^2.2.2",
     "babel": "6.23.0",
     "babel-core": "6.26.0",
@@ -105,7 +113,7 @@
     "eslint-plugin-prettier": "^3.0.1",
     "husky": "^0.14.3",
     "istanbul": "^0.4.3",
-    "jest": "^24.0.0",
+    "jest": "^24.9.0",
     "mocha": "^5.0.1",
     "mocha-lcov-reporter": "^1.2.0",
     "nodemon": "^1.18.7",
@@ -115,6 +123,7 @@
     "sequelize-cli": "^4.0.0",
     "supertest": "^3.4.2",
     "ts-node-dev": "1.0.0-pre.40",
+    "ts-jest": "^24.0.2",
     "typescript": "^3.5.2"
   }
 }
